api-python/pyproject.toml
[project]
name = "akson-api"
version = "0.1.0"
description = "Personal Assistant"
authors = [
    {name = "Cenk Altı", email = "cenkalti@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "sse-starlette>=2.1.3",
    "putio-py>=8.8.0",
    "httpx>=0.27.0",
    "docstring-parser>=0.16",
    "jinja2>=3.1.5",
    "fastapi[standard]>=0.115.8",
    "rich>=13.9.4",
    "mcp>=1.3.0",
    "openai>=1.70.0",
    "litellm>=1.69.0",
    "uvicorn>=0.34.0",
    "langfuse>=2.60.5",
    "fastmcp>=2.6.1",
    "nanoid>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
docs = [
    "griffe-pydantic>=1.1.4",
    "mkdocs-material>=9.6.12",
    "mkdocstrings-python>=1.16.10",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
